requirements.txt
fastapi
fastapi-utils
passlib
password-strength
PyJWT
pymongo
python-dotenv
requests
typing-inspect
pydantic[email]